requirements.txt
PyQt5==5.15.9
SpeechRecognition==3.10.0
google-generativeai==0.3.1
pyaudio
sounddevice
scipy
numpy
setuptools
vosk
python-dotenv
pyttsx3
groq
faiss-cpu
tenacity
requests
tqdm
langchain-google-genai
langchain-community
langchain

# sentence-transformers 
# torch
# gTTS
# playsound
# pydub 
# simpleaudio


# following needs to be installed separately with apt
# espeak
# portaudio19-dev
